new changes

+++ bookings/src/main/java/vttp2023/batch3/assessment/paf/bookings/services/ListingsService.java
@@ -26,7 +26,7 @@ public class ListingsService {
 				.map(d -> new User(
 					d.getString("country"),
 					d.getInteger("numberOfPerson"),
-					d.getFloat("priceRange")
+					d.getInteger("priceRange")
 				))
 			.toList(); 
 	}
